Sharpened the rules a bit to make it more clear.
Added one about new blogs, I've seen a couple being submitted with the exact three posts that are needed. To prevent blogs being added, that are started and never looked at ever again, there is now a rule that it has to be active for at least a year. But maybe that is a bit much? What do you guys think?