Implemented filtering of language
As discussed in #94 we've been wanting the ability to filter on certain languages for some time now. With this change, we have!

We have introduced a new property on the `IAmACommunityMember` interface. This needs to have a ISO 639-1 formatted language code.
All existing authors have been checked and their language code is filled in.

Also implemented a unit test to cover this.